RAG_Backend/src/requirements.txt
ipython
google-generativeai
qdrant-client
fastapi
uvicorn[standard]
beautifulsoup4
numpy
jinja2